% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopviz.R
\name{loopviz}
\alias{loopviz}
\title{Visualize Nested Loops}
\usage{
loopviz(..., animate = FALSE, delay = NULL, loop = TRUE)
}
\arguments{
\item{...}{A vector of integers}
}
\description{
Simply pass a vector of integers. The first integer represents
the number of iterations for the top level loop. The second integer
represents the number of iterations for the first nested loop and so on. As
an example, \code{loopviz(1, 2)} is a loop that iterates once and contains a
nested loop which iterates twice.
}
